[Fix] #171 - FeedRepository와 FeedDetailView 데이터 바인딩 진행 중

+++ WSSiOS/WSSiOS/Source/Data/DTO/FeedDetailResult.swift
@@ -13,7 +13,7 @@ struct Feed {
     var createdDate: String
     var content: String
     var novelTitle: String
-    var novelRating: Double
+    var novelRating: Float
     var novelRatingCount: Int
     var genres: [ReplaceGenre]
     var likeCount: Int
